+++ ydb/core/tx/locks/ya.make
@@ -11,6 +11,7 @@ SRCS(
 PEERDIR(
     ydb/core/protos
     ydb/core/tablet_flat
+    ydb/library/range_treap
 )
 
 YQL_LAST_ABI_VERSION()
