+++ ydb/library/range_treap/ya.make
@@ -1,0 +1,12 @@
+LIBRARY()
+
+SRCS(
+    range_treap.cpp
+)
+
+PEERDIR(
+    util
+    ydb/library/accessor
+)
+
+END()
